requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
boto3==1.35.29
Pillow==10.4.0
Werkzeug==3.1.3
alembic==1.15.2
blinker==1.9.0
click==8.1.8
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
SQLAlchemy==2.0.40
typing_extensions==4.13.2